requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
djoser
docker==7.1.0
drf-yasg==1.21.10
flake8==6.0.0
flake8-isort==6.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.2.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
pytest-docker==3.2.1
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==79.0.1
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
